Fixed hide/resetView. Renamed app.js to server.js. Added install and uninstall to Syslog.
