+++ controllers/networks.js
@@ -1,4 +1,4 @@
-import { docker } from '../app.js';
+import { docker } from '../server.js';
 
 
 export const Networks = async function(req, res) {
